+++ packages/terra-form-select/src/multiple/Frame.jsx
@@ -207,8 +207,7 @@ class Frame extends React.Component {
                 </li>
               </ul>
             </li>
-          ) : null
-        }
+          ) : null}
         <li className={cx('search-wrapper')}>
           <input {...inputAttrs} value={searchValue} />
         </li>
@@ -660,8 +659,7 @@ class Frame extends React.Component {
               {children}
             </Menu>
           </Dropdown>
-          )
-        }
+          )}
       </div>
     );
   }